stockprj/requirements.txt
amqp==5.0.6
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.0
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.3
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-richtextfield==1.6.1
djangorestframework==3.12.4
kombu==5.1.0
Pillow==8.2.0
prompt-toolkit==3.0.18
pycparser==2.20
pytz==2021.1
rabbitmq==0.2.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
